# perf: reuse intermediate alloc