Improve throughput of the following operations by 12% to 33%.

Improve allocations of the following operations by 49% to 76%.

```
▶ benchstat old.txt new.txt
name                                      old time/op    new time/op    delta
/OneNode/128_keys_RueidisDoMulti-10         39.0µs ± 0%    39.0µs ± 0%     ~     (p=0.346 n=10+9)
/OneNode/128_keys_RueidisMGet-10            20.0µs ± 1%    19.9µs ± 1%   -0.52%  (p=0.029 n=10+10)
/OneNode/128_keys_RueidisMGetCache-10       15.6µs ± 2%    11.7µs ± 1%  -24.65%  (p=0.000 n=10+10)
/OneNode/128_keys_RueidisDoMultiCache-10    10.0µs ± 3%     7.8µs ± 1%  -21.88%  (p=0.000 n=10+10)
/Cluster/128_keys_RueidisDoMulti-10         40.8µs ± 4%    35.1µs ± 3%  -13.81%  (p=0.000 n=10+10)
/Cluster/128_keys_RueidisMGet-10            43.6µs ± 3%    38.3µs ± 5%  -12.14%  (p=0.000 n=10+10)
/Cluster/128_keys_RueidisMGetCache-10       16.5µs ± 3%    10.9µs ± 1%  -33.67%  (p=0.000 n=10+9)
/Cluster/128_keys_RueidisDoMultiCache-10    12.1µs ± 1%     8.6µs ± 2%  -28.74%  (p=0.000 n=10+10)

name                                      old alloc/op   new alloc/op   delta
/OneNode/128_keys_RueidisDoMulti-10         18.5kB ± 0%    18.5kB ± 0%     ~     (p=0.000 n=10+10)
/OneNode/128_keys_RueidisMGet-10            41.1kB ± 0%    41.1kB ± 0%     ~     (p=0.327 n=10+10)
/OneNode/128_keys_RueidisMGetCache-10       60.8kB ± 0%    30.8kB ± 1%  -49.36%  (p=0.000 n=9+9)
/OneNode/128_keys_RueidisDoMultiCache-10    32.7kB ± 0%    11.6kB ± 1%  -64.37%  (p=0.000 n=8+10)
/Cluster/128_keys_RueidisDoMulti-10         48.3kB ± 0%    19.4kB ± 0%  -59.87%  (p=0.000 n=8+10)
/Cluster/128_keys_RueidisMGet-10            75.1kB ± 0%    36.3kB ± 1%  -51.63%  (p=0.000 n=8+10)
/Cluster/128_keys_RueidisMGetCache-10       76.7kB ± 0%    30.6kB ± 1%  -60.13%  (p=0.000 n=10+10)
/Cluster/128_keys_RueidisDoMultiCache-10    48.8kB ± 0%    11.6kB ± 2%  -76.30%  (p=0.000 n=10+10)

name                                      old allocs/op  new allocs/op  delta
/OneNode/128_keys_RueidisDoMulti-10            130 ± 0%       130 ± 0%     ~     (all equal)
/OneNode/128_keys_RueidisMGet-10               133 ± 0%       133 ± 0%     ~     (all equal)
/OneNode/128_keys_RueidisMGetCache-10         37.0 ± 0%      20.0 ± 0%  -45.95%  (p=0.000 n=10+10)
/OneNode/128_keys_RueidisDoMultiCache-10      33.0 ± 0%      17.0 ± 0%  -48.48%  (p=0.000 n=10+8)
/Cluster/128_keys_RueidisDoMulti-10            176 ± 0%       141 ± 0%  -19.89%  (p=0.000 n=10+10)
/Cluster/128_keys_RueidisMGet-10               181 ± 0%       144 ± 0%  -20.44%  (p=0.000 n=10+10)
/Cluster/128_keys_RueidisMGetCache-10         57.0 ± 0%      20.0 ± 0%  -64.91%  (p=0.000 n=10+10)
/Cluster/128_keys_RueidisDoMultiCache-10      52.0 ± 0%      16.0 ± 0%  -69.23%  (p=0.000 n=10+10)
```

Benchmark source:
```go
package sep

import (
	"context"
	"math/rand"
	"strconv"
	"testing"
	"time"

	"github.com/redis/rueidis"
)

var charset = []byte("abcdefghijklmnopqrstuvwxyz")

func randstr(n int) string {
	b := make([]byte, n)
	for i := range b {
		b[i] = charset[rand.Intn(len(charset))]
	}
	return string(b)
}

func Benchmark(b *testing.B) {
	testfn := func(b *testing.B, n int, addresses []string) {
		ns := strconv.Itoa(n)
		makeclient := func(addresses []string) rueidis.Client {
			client, err := rueidis.NewClient(rueidis.ClientOption{
				InitAddress: addresses,
			})
			if err != nil {
				panic(err)
			}
			return client
		}

		// prepare keys
		keys := make([]string, n)
		cmds := make(rueidis.Commands, n)

		{
			client := makeclient(addresses)
			for i := range keys {
				keys[i] = randstr(10)
				cmds[i] = client.B().Set().Key(keys[i]).Value(randstr(50)).Build()
			}
			client.Do(context.Background(), client.B().Flushall().Build())
			resps := client.DoMulti(context.Background(), cmds...)
			if len(resps) != len(keys) {
				panic("worn")
			}
			client.Close()
		}
		b.Run(ns+" RueidisDoMulti", func(b *testing.B) {
			client := makeclient(addresses)
			defer client.Close()
			commands := make([]rueidis.Completed, len(keys))
			for i := range commands {
				commands[i] = client.B().Get().Key(keys[i]).Build().Pin()
			}
			b.ReportAllocs()
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					ret := client.DoMulti(context.Background(), commands...)
					if len(ret) != len(keys) {
						panic("wrong")
					}
				}
			})
			b.StopTimer()
		})
		b.Run(ns+" RueidisMGet", func(b *testing.B) {
			client := makeclient(addresses)
			defer client.Close()
			b.ReportAllocs()
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					ret, err := rueidis.MGet(client, context.Background(), keys)
					if err != nil || len(ret) != len(keys) {
						panic(err)
					}
				}
			})
			b.StopTimer()
		})
		b.Run(ns+" RueidisMGetCache", func(b *testing.B) {
			client := makeclient(addresses)
			defer client.Close()
			b.ReportAllocs()
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					ret, err := rueidis.MGetCache(client, context.Background(), time.Minute, keys)
					if err != nil || len(ret) != len(keys) {
						panic(err)
					}
				}
			})
			b.StopTimer()
		})
		b.Run(ns+" RueidisDoMultiCache", func(b *testing.B) {
			client := makeclient(addresses)
			defer client.Close()
			commands := make([]rueidis.CacheableTTL, len(keys))
			for i := range commands {
				commands[i] = rueidis.CT(client.B().Get().Key(keys[i]).Cache().Pin(), time.Minute)
			}
			b.ReportAllocs()
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					ret := client.DoMultiCache(context.Background(), commands...)
					if len(ret) != len(keys) {
						panic("wrong")
					}
				}
			})
			b.StopTimer()
		})
	}

	b.Run("OneNode", func(b *testing.B) {
		for _, n := range []int{128} {
			testfn(b, n, []string{"127.0.0.1:6379"})
		}
	})

	b.Run("Cluster", func(b *testing.B) {
		for _, n := range []int{128} {
			testfn(b, n, []string{"127.0.0.1:7001", "127.0.0.1:7002", "127.0.0.1:7003"})
		}
	})
}
```